fix: use uint128 for upper and lower threshold

+++ src/asset-management/AaveManager.sol
@@ -29,9 +29,8 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
     /// @dev percentage of the pool's assets, above and below which
     /// the manager will divest the shortfall and invest the excess
     /// 1e18 == 100%
-    // TODO: can we use uint128 for each of these for gas savings?
-    uint256 public upperThreshold = 0.7e18; // 70%
-    uint256 public lowerThreshold = 0.3e18; // 30%
+    uint128 public upperThreshold = 0.7e18; // 70%
+    uint128 public lowerThreshold = 0.3e18; // 30%
 
     /// @dev this contract itself is immutable and is the source of truth for all relevant addresses for aave
     IPoolAddressesProvider public immutable addressesProvider;
@@ -91,12 +90,12 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
         windDownMode = aWindDown;
     }
 
-    function setUpperThreshold(uint256 aUpperThreshold) external onlyOwner {
+    function setUpperThreshold(uint128 aUpperThreshold) external onlyOwner {
         require(aUpperThreshold <= 1e18 && aUpperThreshold >= lowerThreshold, "AM: INVALID_THRESHOLD");
         upperThreshold = aUpperThreshold;
     }
 
-    function setLowerThreshold(uint256 aLowerThreshold) external onlyOwner {
+    function setLowerThreshold(uint128 aLowerThreshold) external onlyOwner {
         require(aLowerThreshold <= 1e18 && aLowerThreshold <= upperThreshold, "AM: INVALID_THRESHOLD");
         lowerThreshold = aLowerThreshold;
     }
@@ -276,10 +275,10 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
     function _calculateChangeAmount(uint256 aReserve, uint256 aManaged) internal view returns (int256 rAmountChange) {
         uint256 lRatio = aManaged.divWad(aReserve);
         if (lRatio < lowerThreshold) {
-            rAmountChange = int256(aReserve.mulWad(lowerThreshold.avg(upperThreshold)) - aManaged);
+            rAmountChange = int256(aReserve.mulWad(uint256(lowerThreshold).avg(upperThreshold)) - aManaged);
             assert(rAmountChange > 0);
         } else if (lRatio > upperThreshold) {
-            rAmountChange = int256(aReserve.mulWad(lowerThreshold.avg(upperThreshold))) - int256(aManaged);
+            rAmountChange = int256(aReserve.mulWad(uint256(lowerThreshold).avg(upperThreshold))) - int256(aManaged);
             assert(rAmountChange < 0);
         }
     }
